+++ app/Config/routes.php
@@ -136,12 +136,21 @@
 	Router::connect('/users/index/*', array('controller' => 'users', 'action' => 'index'));
 	Router::connect('/register', array('controller' => 'users', 'action' => 'register'));
 	Router::connect('/register/thankyou', array('controller' => 'users', 'action' => 'register_thankyou'));
+	Router::connect('/recover', array('controller' => 'users', 'action' => 'recover'));
+	Router::connect('/recover/thankyou', array('controller' => 'users', 'action' => 'recover_thankyou'));
 	Router::connect('/users/*', array('controller' => 'users', 'action' => 'delete', '[method]' => 'DELETE'));
 	Router::connect('/users/password/*', array('controller' => 'users', 'action' => 'editPassword', '[method]' => 'PUT'));
 	Router::connect('/users/avatar/*', array('controller' => 'users', 'action' => 'editAvatar', '[method]' => 'POST'));
 	Router::connect('/users/validateAccount/*', array('controller' => 'users', 'action' => 'validateAccount'));
 	Router::connect('/users/*', array('controller' => 'users', 'action' => 'edit', '[method]' => 'PUT'));
 	Router::connect('/users/*', array('controller' => 'users', 'action' => 'view', '[method]' => 'GET'));
+
+/**
+ * Custom route for setup
+ */
+	Router::connect('/setup/recover', array('controller' => 'setup', 'action' => 'recover', '[method]' => 'GET'));
+	Router::connect('/setup/completeRecovery', array('controller' => 'setup', 'action' => 'completeRecovery', '[method]' => 'PUT'));
+
 
 /**
  * Custom route for permissions controller