+++ app/webroot/js/app/component/resource_shortcuts.js
@@ -17,7 +17,7 @@ import 'mad/component/menu';
  */
 var ResourceShortcuts = passbolt.component.ResourceShortcuts = mad.component.Menu.extend('passbolt.component.ResourceShortcuts', /** @static */ {
 
-	'defaults': {}
+	defaults: {}
 
 }, /** @prototype */ {
 
@@ -28,61 +28,66 @@ var ResourceShortcuts = passbolt.component.ResourceShortcuts = mad.component.Men
 	afterStart: function() {
 		var menuItems = [
 			new mad.model.Action({
-				'id': 'js_pwd_wsp_filter_all',
-				'label': __('All items'),
-				'action': function () {
+				id: 'js_pwd_wsp_filter_all',
+				label: __('All items'),
+				case : 'all_items',
+				action: function () {
 					var filter = new passbolt.model.Filter({
-						'label': __('All items'),
-						'order': 'modified',
-                        'case' : 'all_items',
-						'type': passbolt.model.Filter.SHORTCUT
+						label: __('All items'),
+						order: 'modified',
+                        case : 'all_items',
+						type: passbolt.model.Filter.SHORTCUT
 					});
-					mad.bus.trigger('filter_resources_browser', filter);
+					mad.bus.trigger('filter_workspace', filter);
 				}
 			}), new mad.model.Action({
-				'id': 'js_pwd_wsp_filter_favorite',
-				'label': __('Favorite'),
-				'action': function () {
+				id: 'js_pwd_wsp_filter_favorite',
+				label: __('Favorite'),
+				case: 'favorite',
+				action: function () {
 					var filter = new passbolt.model.Filter({
-						'label': __('Favorite'),
-						'case': 'favorite',
-						'type': passbolt.model.Filter.SHORTCUT
+						label: __('Favorite'),
+						case: 'favorite',
+						type: passbolt.model.Filter.SHORTCUT
 					});
-					mad.bus.trigger('filter_resources_browser', filter);
+					mad.bus.trigger('filter_workspace', filter);
 				}
 			}), new mad.model.Action({
-				'id': 'js_pwd_wsp_filter_modified',
-				'label': __('Recently modified'),
-				'action': function () {
+				id: 'js_pwd_wsp_filter_modified',
+				label: __('Recently modified'),
+				case : 'recently_modified',
+				action: function () {
 					var filter = new passbolt.model.Filter({
-						'label': __('Recently modified'),
-						'order': 'modified',
-                        'case' : 'recently_modified',
-						'type': passbolt.model.Filter.SHORTCUT
+						label: __('Recently modified'),
+						order: 'modified',
+                        case : 'recently_modified',
+						type: passbolt.model.Filter.SHORTCUT
 					});
-					mad.bus.trigger('filter_resources_browser', filter);
+					mad.bus.trigger('filter_workspace', filter);
 				}
 			}), new mad.model.Action({
-				'id': 'js_pwd_wsp_filter_share',
-				'label': __('Shared with me'),
-				'action': function () {
+				id: 'js_pwd_wsp_filter_share',
+				label: __('Shared with me'),
+				case: 'shared',
+				action: function () {
 					var filter = new passbolt.model.Filter({
-						'label': __('Shared with me'),
-						'case': 'shared',
-						'type': passbolt.model.Filter.SHORTCUT
+						label: __('Shared with me'),
+						case: 'shared',
+						type: passbolt.model.Filter.SHORTCUT
 					});
-					mad.bus.trigger('filter_resources_browser', filter);
+					mad.bus.trigger('filter_workspace', filter);
 				}
 			}), new mad.model.Action({
-				'id': 'js_pwd_wsp_filter_own',
-				'label': __('Items I own'),
-				'action': function () {
+				id: 'js_pwd_wsp_filter_own',
+				label: __('Items I own'),
+				case: 'own',
+				action: function () {
 					var filter = new passbolt.model.Filter({
-						'label': __('Items I own'),
-						'case': 'own',
-						'type': passbolt.model.Filter.SHORTCUT
+						label: __('Items I own'),
+						case: 'own',
+						type: passbolt.model.Filter.SHORTCUT
 					});
-					mad.bus.trigger('filter_resources_browser', filter);
+					mad.bus.trigger('filter_workspace', filter);
 				}
 			})
 		];
@@ -96,14 +101,23 @@ var ResourceShortcuts = passbolt.component.ResourceShortcuts = mad.component.Men
 	/* ************************************************************** */
 
 	/**
-	 * Listen to the browser filter
+	 * Listen to the workspace is filtered
 	 * @param {jQuery} element The source element
 	 * @param {Event} event The jQuery event
 	 * @param {passbolt.model.Filter} filter The filter to apply
 	 */
-	'{mad.bus.element} filter_resources_browser': function (element, evt, filter) {
+	'{mad.bus.element} filter_workspace': function (element, evt, filter) {
+		var self = this;
+
 		if (filter.type != passbolt.model.Filter.SHORTCUT) {
 			this.unselectAll();
+		} else {
+			this.options.items.each(function(item, i) {
+				if (item.case == filter.case) {
+					self.selectItem(item);
+					return;
+				}
+			});
 		}
 	}
 });