+++ app/webroot/js/app/component/user_browser.js
@@ -6,7 +6,7 @@ import 'app/model/group';
 import 'app/model/profile';
 import 'app/view/component/user_browser';
 
-/*
+/**
  * @class passbolt.component.UserBrowserController
  * @inherits {mad.controller.component.GridController}
  * @parent index
@@ -19,7 +19,7 @@ import 'app/view/component/user_browser';
  * @param {HTMLElement} element the element this instance operates on.
  * @param {Object} [options] option values for the controller.  These get added to
  * this.options and merged with defaults static variable
- * @return {passbolt.component.UserBrowserController}
+ * @return {passbolt.component.UserBrowser}
  */
 var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('passbolt.component.UserBrowser', /** @static */ {
 
@@ -28,8 +28,6 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
         itemClass: passbolt.model.User,
         // Specific view for userBrowser. To handle specific behaviours like drag n drop.
         viewClass: passbolt.view.component.UserBrowser,
-        // the list of resources displayed by the grid
-        users: new can.Model.List(),
         // the list of displayed categories
         // categories: new passbolt.model.Category.List()
         groups: [],
@@ -43,30 +41,46 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
 
 }, /** @prototype */ {
 
+    /**
+     * The filter used to filter the browser.
+     * @type {passbolt.model.Filter}
+     */
+    filterSettings: null,
+
+    /**
+     * Keep a trace of the old filter used to filter the browser.
+     * @type {passbolt.model.Filter}
+     */
+    oldFilterSettings: null,
+
     // Constructor like
     init: function (el, options) {
 
         // The map to use to make our grid working with our resource model
         options.map = new mad.Map({
             id: 'id',
-            name: 'name',
+            name: {
+                key: 'Profile',
+                func: function(profile) {
+                    return profile.first_name + ' ' + profile.last_name;
+                }
+            },
             username: 'username',
             modified: 'modified',
+            last_logged_in: 'last_logged_in',
             Group: 'Group',
             Profile: 'Profile'
         });
 
         // the columns model
-        options.columnModel = [{
+        options.columnModel = [new mad.model.GridColumn({
             name: 'multipleSelect',
             index: 'multipleSelect',
-            header: {
-                css: ['selections s-cell'],
-                label: '<div class="input checkbox">'
-                + '<input type="checkbox" name="select all" value="checkbox-select-all" id="checkbox-select-all-people" disabled="disabled">'
-                + '<label for="checkbox-select-all-people">select all</label> \
-                                    </div>'
-            },
+            css: ['selections s-cell'],
+            label: '<div class="input checkbox"> \
+                    <input type="checkbox" name="select all" value="checkbox-select-all" id="checkbox-select-all-people" disabled="disabled"> \
+                    <label for="checkbox-select-all-people">select all</label> \
+                </div>',
             cellAdapter: function (cellElement, cellValue, mappedItem, item, columnModel) {
                 var availableValues = [];
                 availableValues[item.id] = '';
@@ -82,49 +96,75 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
                 );
                 checkbox.start();
             }
-        }, {
+        }), new mad.model.GridColumn({
             name: 'avatar',
             index: 'Profile',
-            header: {
-                css: ['s-cell'],
-                label: ''
-            },
+            css: ['s-cell'],
+            label: '',
             titleAdapter: function (value, mappedItem, item, columnModel) {
                 return 'avatar';
             },
             valueAdapter: function (value, mappedItem, item, columnModel) {
                 return '<img src="' + item.Profile.avatarPath('small') + '" alt="' + __('Picture of: ') + mappedItem.Profile.first_name + ' ' + mappedItem.Profile.last_name + '" width="30" height="30">';
             }
-        }, {
+        }), new mad.model.GridColumn({
             name: 'name',
             index: 'Profile',
-            header: {
-                css: ['m-cell'],
-                label: __('User')
-            },
-            valueAdapter: function (value, mappedItem, item, columnModel) {
-                return mappedItem.Profile.first_name + ' ' + mappedItem.Profile.last_name;
-            }
-        }, {
+            css: ['m-cell'],
+            label: __('User'),
+            sortable: true
+        }), new mad.model.GridColumn({
             name: 'username',
             index: 'username',
-            header: {
-                css: ['m-cell'],
-                label: __('Username')
-            }
-        }, {
+            css: ['m-cell'],
+            label: __('Username'),
+            sortable: true
+        }), new mad.model.GridColumn({
             name: 'modified',
             index: 'modified',
-            header: {
-                css: ['m-cell'],
-                label: __('Modified')
-            },
+            css: ['m-cell'],
+            label: __('Modified'),
+            sortable: true,
             valueAdapter: function (value, mappedItem, item, columnModel) {
 	            return passbolt.Common.datetimeGetTimeAgo(value);
             }
-        }];
+        }), new mad.model.GridColumn({
+            name: 'last_logged_in',
+            index: 'last_logged_in',
+            css: ['m-cell'],
+            label: __('Last logged in'),
+            sortable: true,
+            valueAdapter: function (value, mappedItem, item, columnModel) {
+                var last_logged_in = __('never');
+                if (value != undefined) {
+                    last_logged_in = passbolt.Common.datetimeGetTimeAgo(value);
+                }
+                return last_logged_in;
+            }
+        })];
 
         this._super(el, options);
+    },
+
+    /**
+     * Get a target column model of the grid.
+     * If no target
+     *
+     * @todo move this function to the parent class mad.grid
+     * @return {mad.model.Model}
+     */
+    getColumnModel: function (name) {
+        var returnValue = null;
+        if (name != undefined) {
+            for (var i in this.options.columnModel) {
+                if (this.options.columnModel[i].name == name) {
+                    return this.options.columnModel[i];
+                }
+            }
+        } else {
+            returnValue = this.options.columnModel;
+        }
+        return returnValue;
     },
 
     /**
@@ -144,7 +184,6 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
 
         // Is the selected user same as the current user.
         var isSelf = passbolt.model.User.getCurrent().id == this.options.selectedUsers[0].id;
-
 
         // Instantiate the contextual menu menu.
         var contextualMenu = new mad.component.ContextualMenu(null, {
@@ -171,6 +210,7 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
             }
         });
         contextualMenu.insertItem(action);
+
         // Add Edit action.
         var action = new mad.model.Action({
             id: 'js_user_browser_menu_copy_email',
@@ -220,27 +260,16 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
     },
 
     /**
-     * Insert a resource in the grid
-     * @param {mad.model.Model} resource The resource to insert
-     * @param {string} refResourceId The reference resource id. By default the grid view object
-     * will choose the root as reference element.
-     * @param {string} position The position of the newly created item. You can pass in one
-     * of those strings: "before", "after", "inside", "first", "last". By dhe default value
-     * is set to last.
-     */
-    insertItem: function (user, refUserId, position) {
-        // add the resource to the list of observed resources
-        this.options.users.push(user);
-        // insert the item to the grid
-        this._super(user, refUserId, position);
-    },
-
-    /**
      * Refresh an item in the grid.
      * We override this function, so we can keep the selected state after the refresh.
      * @param item
      */
     refreshItem: function (item) {
+        // If the item doesn't exist
+        if (!this.itemExists(item)) {
+            return;
+        }
+
         this._super(item);
         if (this.options.selectedUsers.length > 0) {
             this.select(this.options.selectedUsers[0]);
@@ -254,16 +283,6 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
     removeItem: function (item) {
         // remove the item to the grid
         this._super(item);
-    },
-
-    /**
-     * reset
-     */
-    reset: function () {
-        // reset the list of observed resources
-        // by removing a resource from the resources list stored in options, the Browser will
-        // update itself (check "{resources} remove" listener)
-        this.options.users.splice(0, this.options.users.length);
     },
 
     /**
@@ -307,7 +326,12 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
      * @param {boolean} silent Do not propagate any event (default:false)
      */
     select: function (item, silent) {
-        silent = typeof silent == 'undefined' ? false : silent;
+        silent = silent || false;
+
+        // If the item doesn't exist
+        if (!this.itemExists(item)) {
+            return;
+        }
 
         // Unselect the previously selected user, if not in multipleSelection.
         if (!this.state.is('multipleSelection') &&
@@ -348,6 +372,11 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
     unselect: function (item, silent) {
         silent = typeof silent == 'undefined' ? false : silent;
 
+        // If the item doesn't exist
+        if (!this.itemExists(item)) {
+            return;
+        }
+
         // Uncheck the associated checkbox (if it is not already done).
 		var controlId = 'multiple_select_checkbox_' + item.id,
         	checkbox = mad.getControl(controlId, 'mad.form.Checkbox');
@@ -365,6 +394,122 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
         if (!silent) {
             mad.bus.trigger('user_unselected', item);
         }
+    },
+
+    /**
+     * Filter the browser using a filter settings object
+     * @param {passbolt.model.Filter} filter The filter to
+     */
+    filterBySettings: function(filter) {
+        var self = this,
+        // The deferred used for the users find all request.
+            def = $.Deferred();
+
+        // Save the old filter settings.
+        this.oldFilterSettings = this.filterSettings;
+        // Clone the given filter to avoid any changes problem.
+        this.filterSettings = filter.clone();
+
+        //// reset the state variables
+        //this.options.groups = [];
+		//
+        //// override the current list of users displayed with the new ones
+        //var filteredGroup = filter.getForeignModels('Group');
+        //if (filteredGroup) {
+        //    can.each(filteredGroup, function (group, i) {
+        //        self.options.groups.push(group.id);
+        //    });
+        //}
+
+        // If the current filter case is different than the previous filter case
+        //   or this is the initial filtering (loading)
+        if (this.oldFilterSettings == null
+            || this.filterSettings.case != this.oldFilterSettings.case) {
+
+            // Mark the component as loading.
+            // Complete it once the users are retrieved and rendered.
+            this.setState('loading');
+
+            // Remove all elements from the grid
+            this.reset();
+
+            // Retrieve the resources.
+            passbolt.model.User.findAll({
+                filter: this.filterSettings,
+                recursive: true,
+                silentLoading: false
+            }).then(function (users, response, request) {
+                // If the browser has been destroyed before the request completed.
+                if (self.element == null) return;
+
+                // If the grid was marked as filtered, reset it.
+                self.filtered = false;
+
+                // Load the resources in the browser.
+                self.load(users);
+                self.setState('ready');
+                if (!users.length) {
+                    self.state.addState('empty');
+                }
+
+                // @todo should be cleaned with the filter refactoring PASSBOLT-1571
+                if (filter.case == 'all_items') {
+                    var sortedColumnModel = self.getColumnModel('name');
+                    self.view.markColumnAsSorted(sortedColumnModel, true);
+                }
+
+                // If the resources are ordered.
+                if (filter.order != undefined) {
+                    var sortedColumnModel = self.getColumnModel(filter.order);
+                    if (sortedColumnModel != null) {
+                        var orderAsc = true;
+                        self.view.markColumnAsSorted(sortedColumnModel, orderAsc);
+                    }
+                }
+
+                def.resolve();
+            });
+        } else {
+            def.resolve();
+        }
+
+        // When the resources have been retrieved.
+        $.when(def).done(function() {
+            // Filter by keywords.
+            var keywords = filter.getKeywords();
+            if (keywords != '') {
+                self.filterByKeywords(keywords, {
+                    searchInFields: ['username', 'Role.name', 'Profile.first_name', 'Profile.last_name']
+                });
+            } else if (self.isFiltered()){
+                self.resetFilter();
+            }
+        });
+
+        return def;
+    },
+
+    /**
+     * Does the item exist
+     * @param {passbolt.Model} item The item to check if it existing
+     * @return {boolean}
+     * @todo PASSBOLT-1614 move this function into mad grid.
+     */
+    itemExists: function (item) {
+        return this.view.getItemElement(item).length > 0 ? true : false;
+    },
+
+    /**
+     * Reset the filtering
+     * @todo PASSBOLT-1614 move this function into mad grid.
+     */
+    resetFilter: function () {
+        var self = this;
+        this.options.isFiltered = false;
+
+        can.each(this.options.items, function(item, i) {
+            self.view.showItem(item);
+        });
     },
 
     /* ************************************************************** */
@@ -394,7 +539,7 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
      * @param {passbolt.model.User} user The updated user
      */
     '{passbolt.model.User} updated': function (model, ev, user) {
-        if (this.options.users.indexOf(user) != -1) {
+        if (this.options.items.indexOf(user) != -1) {
             this.refreshItem(user);
         }
     },
@@ -422,12 +567,12 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
      */
     '{passbolt.model.GroupUser} destroyed': function (model, ev, groupUser) {
         // Remove user from the list of users in the grid.
-        for (i in this.options.users) {
-            if (this.options.users[i].id == groupUser.user_id) {
+        for (i in this.options.items) {
+            if (this.options.items[i].id == groupUser.user_id) {
                 break;
             }
         }
-        this.options.users.splice(i, 1);
+        this.options.items.splice(i, 1);
 
         // Remove user from the list of selected users.
         for (i in this.options.selectedUsers) {
@@ -491,8 +636,8 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
         //}
 
         // find the resource to select functions of its id
-        var i = mad.model.List.indexOf(this.options.users, userId);
-        var user = this.options.users[i];
+        var i = mad.model.List.indexOf(this.options.items, userId);
+        var user = this.options.items[i];
 
         if (this.beforeSelect(user)) {
             this.select(user);
@@ -510,8 +655,8 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
         var self = this;
 
         // find the resource to select functions of its id
-        var i = mad.model.List.indexOf(this.options.users, userId);
-        var user = this.options.users[i];
+        var i = mad.model.List.indexOf(this.options.items, userId);
+        var user = this.options.items[i];
 
         if (this.beforeUnselect()) {
             self.unselect(user);
@@ -537,39 +682,20 @@ var UserBrowser = passbolt.component.UserBrowser = mad.component.Grid.extend('pa
      * @param {Event} event The jQuery event
      * @param {passbolt.model.Filter} filter The filter to apply
      */
-    '{mad.bus.element} filter_users_browser': function (element, evt, filter) {
-        var self = this;
-        // store the filter
-        this.filter = filter;
-        // reset the state variables
-        this.options.groups = [];
+    '{mad.bus.element} filter_workspace': function (element, evt, filter) {
+        this.filterBySettings(filter);
+    },
 
-        // override the current list of users displayed with the new ones
-        var filteredGroup = filter.getForeignModels('Group');
-        if (filteredGroup) {
-            can.each(filteredGroup, function (group, i) {
-                self.options.groups.push(group.id);
-            });
+    /**
+     * Observe when an item is unselected
+     * @param {HTMLElement} el The element the event occurred on
+     * @param {HTMLEvent} ev The event which occurred
+     * @param {passbolt.model.Resource|array} items The unselected items
+     */
+    '{selectedUsers} remove': function (el, ev, items) {
+        for (var i in items) {
+            this.unselect(items[i]);
         }
-
-        // change the state of the component to loading.
-        this.setState('loading');
-
-        // load resources functions of the filter.
-        passbolt.model.User.findAll({
-            filter: this.filter,
-            recursive: true
-        }, function (users, response, request) {
-            // If the browser component has been destroyed.
-            if (self.element == null) {
-                return;
-            }
-
-            // load the users in the browser.
-            self.load(users);
-            // change the state to ready.
-            self.setState('ready');
-        });
     },
 
     /* ************************************************************** */