+++ app/webroot/js/lib/mad/src/component/grid.js
@@ -1,7 +1,9 @@
 import 'mad/component/component';
+import 'mad/model/grid_column';
 import 'mad/view/component/grid';
 import 'mad/view/template/component/grid/grid.ejs!';
 import 'mad/view/template/component/grid/gridItem.ejs!';
+import 'mad/view/template/component/grid/gridColumnHeader.ejs!';
 
 /**
  * @parent Mad.components_api
@@ -28,8 +30,10 @@ var Grid = mad.component.Grid = mad.Component.extend('mad.component.Grid', {
         viewClass: mad.view.component.Grid,
         // Prefix the id of each row.
         prefixItemId: '',
-        // The Model Class that defines the items displayed by the tree.
+        // The Model Class that defines the items displayed by the grud.
         itemClass: null,
+        // The Map Class used to defined the column model.
+        columnModelClass: mad.model.GridColumn,
         // the grid column model
         columnModel: [],
         // The map used to transform the raw data into expected view format.
@@ -40,16 +44,51 @@ var Grid = mad.component.Grid = mad.Component.extend('mad.component.Grid', {
             item_selected: null,
             // An item is hovered.
             item_hovered: null
-        }
+        },
+        // The items the grid works with.
+        items: new can.Model.List(),
+        // Is the grid filtered.
+        isFiltered: false,
+        // Is the grid sorted.
+        isSorted: false
     }
 
 }, /** @prototype */ {
 
     /**
-     * Flush the grid
+     * Constructor.
+     * Instantiate a new Grid Component.
+     * @param {HTMLElement|can.NodeList|CSSSelectorString} el The element the control will be created on
+     * @param {Object} [options] option values for the component.  These get added to
+     * this.options and merged with defaults static variable
+     * @return {mad.component.Grid}
      */
-    reset: function () {
-        this.view.reset();
+    init: function(el, options) {
+        this._super(el, options);
+        this.options.items = new can.Model.List();
+
+        // Keep a trace of the items after mapping.
+        // This data will be used for post rendering treatments :
+        // * sort ;
+        this.mappedItems = {};
+
+        this.on();
+    },
+
+    /**
+     * After start hook().
+     */
+    afterStart: function() {
+        var columnModel = this.getColumnModel();
+
+        // Associate columnModel definitions to corresponding DOM elements th column header.
+        // It will be used by the view to retrieve associated column model definition.
+        for (var i in columnModel) {
+            var el = $('th.js_grid_column_' + columnModel[i].name, this.element);
+            can.data(el, this.getColumnModelClass().fullName, columnModel[i]);
+        }
+
+        this._super();
     },
 
     /**
@@ -77,6 +116,24 @@ var Grid = mad.component.Grid = mad.Component.extend('mad.component.Grid', {
      */
     getItemClass: function () {
         return this.options.itemClass;
+    },
+
+    /**
+     * Get the column model class.
+     *
+     * @return {can.Map}
+     */
+    getColumnModelClass: function () {
+        return this.options.columnModelClass;
+    },
+
+    /**
+     * Return true if the grid is filtered, else return false.
+     *
+     * @returns {boolean}
+     */
+    isFiltered: function() {
+        return this.options.isFiltered;
     },
 
     /**
@@ -159,6 +216,8 @@ var Grid = mad.component.Grid = mad.Component.extend('mad.component.Grid', {
     removeItem: function (item) {
         // Remove the item to the view
         this.view.removeItem(item);
+        // Free space, remove the relative mapped item
+        delete this.mappedItems[item.id];
     },
 
     /**
@@ -190,8 +249,12 @@ var Grid = mad.component.Grid = mad.Component.extend('mad.component.Grid', {
             throw mad.Exception.get(mad.error.MISSING_OPTION, 'map');
         }
 
+        // Add the item to the list of observed items
+        this.options.items.push(item);
+
         // Map the item.
         mappedItem = this.getMap().mapObject(item);
+        this.mappedItems[item.id] = mappedItem;
 
         // insert the item in the view
         this.view.insertItem(item, refItem, position);
@@ -232,8 +295,12 @@ var Grid = mad.component.Grid = mad.Component.extend('mad.component.Grid', {
         this.view.refreshItem(item);
 
         var self = this,
-            mappedItem = this.getMap().mapObject(item),
+            mappedItem = null,
             columnModels = this.getColumnModel();
+
+        // Map the item.
+        mappedItem = this.getMap().mapObject(item);
+        this.mappedItems[item.id] = mappedItem;
 
         // apply a widget to cells following the columns model
         for (var j in columnModels) {
@@ -263,6 +330,17 @@ var Grid = mad.component.Grid = mad.Component.extend('mad.component.Grid', {
     },
 
     /**
+     * Reset the grid.
+     * Remove all the displayed (and hidden) items.
+     */
+    reset: function () {
+        // reset the list of observed items
+        // by removing an item from the items list stored in options, the grid will
+        // update itself (check "{items} remove" listener)
+        this.options.items.splice(0, this.options.items.length);
+    },
+
+    /**
      * Load items in the grid. If the grid contain items, reset it.
      *
      * @param {mad.model.Model[]} items The array or list of items to insert in the grid
@@ -271,11 +349,152 @@ var Grid = mad.component.Grid = mad.Component.extend('mad.component.Grid', {
         var self = this;
 
         this.reset();
+        this.options.isFiltered = false;
+        this.options.isSorted = false;
+        this.view.markAsUnsorted();
+
         can.each(items, function (item, i) {
             self.insertItem(item);
         });
 
         return this;
+    },
+
+    /**
+     * Filter items in the grid by keywords
+     * @param {string} needle The string to search in the grid
+     */
+    filterByKeywords: function(needle, options) {
+        options = options || {};
+        var self = this,
+            // The fields to look into.
+            searchInFields = [],
+            // The keywords to search.
+            keywords = needle.split(/\s+/),
+            // Filtered resource.
+            filteredItems = new can.List();
+
+        // The fields to look into have been given in options.
+        if (typeof options.searchInFields != 'undefined') {
+            searchInFields = options.searchInFields;
+        } else {
+            searchInFields = this.options.map.getModelTargetFieldsNames();
+        }
+
+        // Search the keywords in the list of items.
+        can.each(this.options.items, function (item, i) {
+            // Foreach keywords.
+            for (var j in keywords) {
+                var found = false,
+                    field = null,
+                    i= 0;
+
+                // Search in the item fields.
+                while(!found && (field = searchInFields[i])) {
+                    // Is the keyword found in the field.
+                    found = can.getObject(field, item)
+                            .toLowerCase()
+                            .indexOf(keywords[j].toLowerCase()) != -1;
+                    i++;
+                }
+
+                // If the keyword hasn't been found in any field.
+                // Search in the next resource.
+                if (!found) {
+                    return;
+                }
+            }
+
+            filteredItems.push(item);
+        });
+
+        // Filter the grid
+        self.filter(filteredItems);
+    },
+
+    /**
+     * Filter items in the grid
+     * @param {can.List} items The list of items to filter
+     */
+    filter: function (filteredItems) {
+        var self = this;
+        this.options.isFiltered = true;
+
+        can.each(this.options.items, function(item, i) {
+            if (filteredItems.indexOf(item) != -1) {
+                self.view.showItem(item);
+            } else {
+                self.view.hideItem(item);
+            }
+        });
+    },
+
+    /**
+     * Sort the grid functions of a given column.
+     * @param columnModel The column the grid should be sort in functions of.
+     * @param sortAsc Should the sort be ascending. True by default.
+     */
+    sort: function (columnModel, sortAsc) {
+        this.options.isSorted = true;
+
+        // Retrieve the mapped item attribute name.
+        var columnId = columnModel.name;
+        // Copy the mappedItems associativate array into array.
+        var mappedItemsCopy = $.map(this.mappedItems, function(value, index) {
+            value.id = index;
+            return [value];
+        });
+
+        // Sort the mapped items
+        mappedItemsCopy.sort(function(itemA, itemB){
+            // ignore upper and lowercase
+            var valueA = itemA[columnId] ? itemA[columnId].toUpperCase() : '',
+                valueB = itemB[columnId] ? itemB[columnId].toUpperCase() : '';
+
+            if (valueA < valueB) {
+                return sortAsc ? -1 : 1;
+            }
+            else if (valueA > valueB) {
+                return sortAsc ? 1 : -1;
+            }
+
+            return 0;
+        });
+
+        // Move all the items following the sort result
+        for (var i in mappedItemsCopy) {
+            this.moveItem(mappedItemsCopy[i], i);
+        }
+
+        // Mark the column as sorted
+        this.view.markColumnAsSorted(columnModel, sortAsc);
+    },
+
+    /**
+     * Move an item to another position in the grid.
+     * @param item The item to move
+     * @param position The position to move the item to
+     */
+    moveItem: function(item, position) {
+        this.view.moveItem(item, position);
+    },
+
+    /* ************************************************************** */
+    /* LISTEN TO THE MODEL EVENTS */
+    /* ************************************************************** */
+
+    /**
+     * Observe when items are removed from the list of observed items and
+     * remove it from the grid
+     * @param {mad.model.Model} model The model reference
+     * @param {HTMLEvent} ev The event which occurred
+     * @param {can.Model.List} items The removed items
+     */
+    '{items} remove': function (model, ev, items) {
+        var self = this;
+        can.each(items, function (item, i) {
+            self.removeItem(item);
+        });
     },
 
     /* ************************************************************** */
@@ -294,6 +513,22 @@ var Grid = mad.component.Grid = mad.Component.extend('mad.component.Grid', {
      */
     'tbody mouseleave': function (element, evt) {
         //
+    },
+
+    /**
+     * @function mad.component.Grid.__column_sort_asc
+     * @parent mad.component.Grid.view_events
+     *
+     * Observe when a sort is requested on a column.
+     *
+     * @param {HTMLElement} el The element the event occurred on
+     * @param {HTMLEvent} ev The event that occurred
+     * @param {mad.model.Gridcolumn} columnModel The column model reference
+     * @param {boolean} sortAsc Should the sort be ascending. If false, the sort will be descending.
+     * @param {HTMLEvent} srcEvent The source event which occurred
+     */
+    ' column_sort': function (el, ev, columnModel, sortAsc, srcEvent) {
+        this.sort(columnModel, sortAsc);
     },
 
     /**